requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.4.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
fastapi==0.70.1
frozenlist==1.2.0
google-api-core==2.3.2
google-auth==2.3.3
google-cloud-core==2.2.1
google-cloud-pubsub==2.9.0
google-cloud-storage==1.44.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
greenlet==1.1.2
grpc-google-iam-v1==0.12.3
grpcio==1.43.0
grpcio-status==1.43.0
h11==0.12.0
idna==3.3
libcst==0.3.23
memory-profiler==0.60.0
multidict==5.2.0
mypy-extensions==0.4.3
proto-plus==1.19.8
protobuf==3.19.1
psutil==5.9.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
python-dotenv==0.19.2
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
slack==0.0.2
slackclient==2.9.3
sniffio==1.2.0
SQLAlchemy==1.4.29
starlette==0.16.0
typing-extensions==4.0.1
typing-inspect==0.7.1
urllib3==1.26.7
uvicorn==0.16.0
yarl==1.7.2
